H1: Change data types to floats
HWP1 and HW1P2, change data types from doubles to floats. Doubles unneded

+++ HW1P1_Code.java
@@ -5,28 +5,28 @@ public class HW1P1_Code {
 
         Scanner A1_input = new Scanner(System.in);
         System.out.println("Please provide your first angle: ");
-        double A1 = A1_input.nextDouble();
+        float A1 = A1_input.nextFloat();
 
         while (A1 <= 0 || A1 >= 180) {
             System.out.println("ERROR: PLease ensure your first angle is a positive number, and is less than 180 degrees.");
             System.out.println("Please provide your first angle: ");
-            A1 = A1_input.nextDouble();
+            A1 = A1_input.nextFloat();
         }
 
         Scanner A2_input = new Scanner(System.in);
         System.out.println("Please provide your second angle: ");
-        double A2 = A2_input.nextDouble();
+        float A2 = A2_input.nextFloat();
 
         while (A2 <= 0 || A2 >= 180) {
             System.out.println("ERROR: Please ensure your second angle is a positive number, and is less than 180 degrees.");
             System.out.println("Please provide your second angle: ");
-            A2 = A2_input.nextDouble();
+            A2 = A2_input.nextFloat();
         }
 
         while (A1 + A2 >= 180) {
             System.out.println("ERROR: Please ensure your first two angles do not add up to 180 degrees or more.");
             System.out.println("Please provide your second angle: ");
-            A2 = A2_input.nextDouble();
+            A2 = A2_input.nextFloat();
         }
 
         A1_input.close();
